CRUD operations functional with login authentication

+++ backend/controllers/leaderController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler');
 
 const Leader = require('../models/leaderModel');
+const User = require('../models/userModel');
 
 //@desc     Get goals
 //@route    GET /api/goals
 //@access   Private
 const getLeaders = asyncHandler(async (req, res) => {
-    const leaders = await Leader.find()
+    const leaders = await Leader.find({ user: req.user.id })
 
     res.status(200).json(leaders)
 });
@@ -19,8 +20,9 @@ const setLeader = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('No streak detected')
     }
-    
-    const leader = await Leader.create( {
+
+    const leader = await Leader.create({
+        user: req.user.id,
         name: req.body.name,
         streak: req.body.streak,
         message: req.body.message
@@ -32,12 +34,26 @@ const setLeader = asyncHandler(async (req, res) => {
 //@desc     Update goals
 //@route    PUT /api/goals/:id
 //@access   Private
-const updateLeader =  asyncHandler(async (req, res) => {
+const updateLeader = asyncHandler(async (req, res) => {
     const leader = await Leader.findById(req.params.id)
 
     if (!leader) {
         res.status(400);
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id);
+
+    //Check for user
+    if (!user) {
+        res.status(401);
+        throw new Error('User not found')
+    }
+
+    //Check if logged in user matches goal user
+    if (global.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User not authorized');
     }
 
     const updatedLeader = await Leader.findByIdAndUpdate(req.params.id, req.body, {
@@ -56,6 +72,20 @@ const deleteLeader = asyncHandler(async (req, res) => {
     if (!leader) {
         res.status(400);
         throw new Error('Leader not found')
+    }
+
+    const user = await User.findById(req.user.id);
+
+    //Check for user
+    if (!user) {
+        res.status(401);
+        throw new Error('User not found')
+    }
+
+    //Check if logged in user matches goal user
+    if (leader.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User not authorized');
     }
 
     await leader.deleteOne()
